Added files via upload
Due to technological difficulties, this couldn't be properly integrated into the Android application at this time. Rather than letting this code go to waste, I feel those who want to know how this program was meant to work should be able to check this code out if they so choose.